ppc64: add asm/sections.h

+++ arch/ppc64/kernel/traps.c
@@ -348,12 +348,12 @@ check_bug_trap(struct pt_regs *regs)
 		/* this is a WARN_ON rather than BUG/BUG_ON */
 		printk(KERN_ERR "Badness in %s at %s:%d\n",
 		       bug->function, bug->file,
-		       bug->line & ~BUG_WARNING_TRAP);
+		      (unsigned int)bug->line & ~BUG_WARNING_TRAP);
 		dump_stack();
 		return 1;
 	}
 	printk(KERN_CRIT "kernel BUG in %s at %s:%d!\n",
-	       bug->function, bug->file, bug->line);
+	       bug->function, bug->file, (unsigned int)bug->line);
 	return 0;
 }
